rename package dots to dtos
