+++ src/main/java/kitchenpos/products/interfaces/ProductApi.java
@@ -6,9 +6,9 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 import javax.validation.Valid;
 import kitchenpos.products.application.ProductFacade;
-import kitchenpos.products.domain.dots.ProductCommand.ChangePriceCommand;
-import kitchenpos.products.domain.dots.ProductCommand.RegisterProductCommand;
-import kitchenpos.products.domain.dots.ProductInfo;
+import kitchenpos.products.domain.dtos.ProductCommand.ChangePriceCommand;
+import kitchenpos.products.domain.dtos.ProductCommand.RegisterProductCommand;
+import kitchenpos.products.domain.dtos.ProductInfo;
 import kitchenpos.products.interfaces.dtos.ProductDto;
 import kitchenpos.products.interfaces.dtos.ProductDto.ProductResponse;
 import org.springframework.http.ResponseEntity;
