+++ src/test/java/kitchenpos/products/application/ProductServiceTest.java
@@ -18,10 +18,10 @@ import kitchenpos.products.domain.ProductRepository;
 import kitchenpos.products.domain.ProductService;
 import kitchenpos.products.domain.ProductServiceImpl;
 import kitchenpos.products.domain.PurgomalumClient;
-import kitchenpos.products.domain.dots.ProductCommand;
-import kitchenpos.products.domain.dots.ProductCommand.ChangePriceCommand;
-import kitchenpos.products.domain.dots.ProductCommand.RegisterProductCommand;
-import kitchenpos.products.domain.dots.ProductInfo;
+import kitchenpos.products.domain.dtos.ProductCommand;
+import kitchenpos.products.domain.dtos.ProductCommand.ChangePriceCommand;
+import kitchenpos.products.domain.dtos.ProductCommand.RegisterProductCommand;
+import kitchenpos.products.domain.dtos.ProductInfo;
 import kitchenpos.products.exception.ProductNameException;
 import kitchenpos.products.exception.ProductPriceException;
 import org.junit.jupiter.api.BeforeEach;
